+++ src/utils/escapeRegExp.js
@@ -1,0 +1,2 @@
+export const escapeRegExp = (string) =>
+    string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
